chore(app): file name and location updates
